Unique repsonse body when auth fails
When authentication fails because the specified domain doesn't exist,
in addition to a 401 respoonse code, return a json response body
indicating the domain doesn't exist so that the client (acorn) can
determine if it needs to request a new domain.

This is to account for the fact that domains eventually expire if they
aren't renewed and acorn will need to know that.