Implement record purging
Add an api endpoint for deleting/purging all the records for a given
domain. This is for the use case when a user disables the acorn-dns
feature. Acorn will detect that config change and call this API endpoint
to immediately purge all records for the cluster's domain.

This will not delete the domain. That will remain in case the user
decides to reenable the feature.