fix: Prevent erroneous purge of wildcard records
We were incorrectly purging wildcard records because of a mistmatched
between what route53 returned and acorn-dns had in its DB. Route53
stores and returns `*` (the wildcard) as `\052`. So, when we did the
comaprison between the record from R53 and the record in the DB, we were
not finding a match and deciding to purge the record for R53.

This fixes the problem by just normalizing the record returned by R53 to
have `*`. Note that R53 accepts `*` just fine and as an equivalent to
`\052`, so when we actually SHOULD delete a wildcard record, this works
as expected.

Signed-off-by: Craig Jellick <craig@acorn.io>